requirements.txt
nicegui
pyjwt
captcha
sqlalchemy
pymysql